+++ src/main/java/com/carspotter/CarSpotter/repository/PenaltyRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PenaltyRepository extends JpaRepository<Penalty, Long> {
+public interface PenaltyRepository extends JpaRepository<Penalty, Integer> {
     List<Penalty> findByIsSecretTrue();
 }
